♻️ refactor(input): move to own directory 'input'
